requirements.txt
openai==1.59.7
pandas==2.2.3
python-dotenv==1.0.1
PyYAML==6.0.2